filewatcher/requirements.txt
python-dotenv
requests
python-dotenv
watchdog